restructure the files
